Split app wallet and walletunlocker services
